requirements.txt
PyQt5
openai~=1.57.0
google-generativeai
python-dotenv